requirements.txt
Click==7.0
Jinja2==2.11.1
livereload==2.6.1
Markdown==3.2.1
MarkupSafe==1.1.1
mkdocs==1.0.4
mkdocstrings==0.7.1
pkg-resources==0.0.0
pymdown-extensions==6.3
PyYAML==5.3
six==1.14.0
tornado==6.0.3
-e git+git@github.com:blogic/udrone.git@758ae32e7fab86d6e20cab21558033e398d0b79b#egg=udronerc
